Revert "feat: support streaming with knowledge base (PL-987) (#799)"
This reverts commit b2629d61859e8725eb34904215dc29ef85c091a2.
